% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_rows.R
\name{drop_rows}
\alias{drop_rows}
\title{drop rows with diagnostic message}
\usage{
drop_rows(df, idx)
}
\arguments{
\item{df}{data.frame}

\item{idx}{logical, TRUE for rows to drop FALSe else}
}
\value{
df after removing rows where idx is TRUE
}
\description{
drop rows with diagnostic message
}
\examples{
drop_rows(df = mtcars, idx = mtcars$cyl==4)
}
