% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_interval.R
\name{in_interval}
\alias{in_interval}
\title{Tag elements inside interval}
\usage{
in_interval(vec, lower, upper, lower_incl, upper_incl, missing_false = FALSE)
}
\arguments{
\item{vec}{numeric}

\item{lower}{numeric, lower bound}

\item{upper}{numeric, upper bound}

\item{lower_incl}{# logical, lower bound inside interval?}

\item{upper_incl}{# logical, upper bound inside interval?}

\item{missing_false}{# logical, if TRUE missing values in vec will be tagged as FALSE}
}
\value{

}
\description{
Tag elements inside interval
}
