% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_duplicates.R
\name{tag_duplicates}
\alias{tag_duplicates}
\title{Tag duplicates and give short diagnostic message}
\usage{
tag_duplicates(df, ...)
}
\arguments{
\item{df}{data.frame to be checked for duplicates}

\item{...}{further options for base::duplicated}
}
\value{
logical, equal to 1 if row is duplicated, 0 else
}
\description{
Tag duplicates and give short diagnostic message
}
\examples{
tag_duplicates(c(9:20, 1:5, 3:7, 0:8))
}
